% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aitchison_dist}
\alias{aitchison_dist}
\title{Compute Aitchison distance between two sets of compositions}
\usage{
aitchison_dist(Pi1, Pi2, V)
}
\arguments{
\item{Pi1, Pi2}{Numeric M x K matrices of compositions.}

\item{V}{Numeric K x (K-1) ILR contrast matrix.}
}
\value{
Numeric vector of length M.
}
\description{
Compute Aitchison distance between two sets of compositions
}
