% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ilr_to_proportions}
\alias{ilr_to_proportions}
\title{Convert ILR scores to topic proportions}
\usage{
ilr_to_proportions(Z, V)
}
\arguments{
\item{Z}{Numeric M x (K-1) matrix of ILR scores.}

\item{V}{Numeric K x (K-1) ILR contrast matrix.}
}
\value{
M x K matrix of topic proportions.
}
\description{
Applies the ILR inverse (softmax of V z) to a matrix of ILR scores.
}
