% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proportions_to_ilr}
\alias{proportions_to_ilr}
\title{Convert topic proportions to ILR scores}
\usage{
proportions_to_ilr(Pi, V)
}
\arguments{
\item{Pi}{Numeric M x K matrix of topic proportions (rows sum to 1,
all entries > 0).}

\item{V}{Numeric K x (K-1) ILR contrast matrix.}
}
\value{
M x (K-1) matrix of ILR scores.
}
\description{
Convert topic proportions to ILR scores
}
